working fps graph

+++ include/Matrix.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <vector>
 
 template <typename T>
@@ -8,9 +9,9 @@ struct Matrix {
     int            sizeX = 0;
     int            sizeY = 0;
 
-    Matrix(int x_, int y_) : v(x_ * y_, T{}), sizeX(x_), sizeY(y_) {}
+    Matrix(int x_, int y_) : v(static_cast<std::size_t>(x_ * y_), T{}), sizeX(x_), sizeY(y_) {}
 
-    T& operator()(int x_, int y_) { return v[x_ + y_ * sizeX]; }
+    T& operator()(int x_, int y_) { return v[static_cast<std::size_t>(x_ + y_ * sizeX)]; }
 
-    const T& operator()(int x_, int y_) const { return v[x_ + y_ * sizeX]; }
+    const T& operator()(int x_, int y_) const { return v[static_cast<std::size_t>(x_ + y_ * sizeX)]; }
 };
